Add Runner and allow out of function call
Add a runner that represent all program-wide information required to
move a state to the next state. If contain a run function that perform
a transition of the machine state-transition system: a full fetch and
execute cycle. The runner is able to load instruction from multiple
functions (but not out of any function yet).

Block now relies on the runner to run code. BB is still independent.