Refactor the representation of registers
Register are now truly plain integers, no more path.
The register index now indexes the path directly,
each register has (in an IdMap):
  - A path: A list of string like ["R8"] or ["PSTATE";"N"]
  - A index: A number identifying the register
  - A type: Bool, Bitvector or Enum

Any register map is now flat, which is much simpler.
To support new register maps, I added "FullVec" a
module of "infinite vector" where all positive integers
have values. In pratice they have a finite vector and a generator.
All values not in the vector are defined by the generator.

I then ported that to Reg.Map where the set of register is not infinite
but may grow when the execution advance. They now contain the same
kind of generator that represent the value taken by not yet added
register. This way all concepts of extending register map (or states)
dissapear.

Register can now be plainly used with Reg.t and the concept of path now
explicitely represents a list of string.

Any check for correct structure of register hierarchy is lost.
The fact that register name have no dots is not check and may
be problematic but those would not be valid Sail identifiers.