Add new caching symbolic memory
This add a new symbolic memory system based on a restricted block
separation. All accesses are bound to block and block cannot alias in
any way. There is a Main block for main memory and other blocks.

Each block is managed with a SymbolicFragment (new module). This
fragments use among other SymbolicBytes as a primitive.

In the idea SymbolicBytes is about storing symbolic values at concrete
addresses whereas symbolicFragment is about storing symbolic values at
symbolic addresses.

The multiple block management code is still in State.Mem.

The Ctype pointer have been modified to hold a provenance field to
represent the pointer provenance i.e. which block it belongs to.
The provenance field is to represent the runtime symbolic
block in which the pointer belong. This is guaranteed
to not alias with any other provenance (Or the main memory).
This is used directly for symbolic execution, and not just for
typing

A breaking change is that in a multi block setup, some memory access
can fail if the type inference system hasn't inferred a provenance.
The only required soundness guarantee from the type system is to not
attribute wrong provenances which is not a big constraint and is
independant of DWARF information.