Add new randomized tests
Reorganize the test structure in tests to have the main
module only about the CLI of the test suite and the other
tests in separate modules.

Switch the tester CLI to Cmdliner to be able to include logs
and config options.

Add randomized tests for concreteEval and Z3 module by testing
one against the other. Fix a lot of bugs in the process,
and implement all the concrete modulus/remainder operations.

In particular I need to add the bvsrem_i class of SMT operations.
They have the same behavior as the one before the i except some
subtelties from Z3.

Also improve the test documentation in the README and the
access to individual test suites in the Makefile.