Make expression annotation be their type (#10)
In State.Exp and Trace.Exp, the annotation is now the SMT type of the
expression. This allow to check the type when in assert mode and
to be able to access that type easily. The replacement is done
when going from Isla ast to the internal ast type which is either in
Trace.of_isla or in IslaRun.

The Exp module is added and provides a lifting of equality, pretty
printing and some other things to expressions.

The ExpTyped module provide the machine typed expression function
and mostly replaces Ast.Op.

All sum manipulation function have been moved to Sums instead of
polutting AstManip and ConcreteEval. The legacy Pointer module
that was originally supposed to do the same thing was scrapped

Since the type of traces changed, any cache is invalidated and the
epoch moves to 3.

The high level interface over Z3 has been entirely refactored is now
functorized over variables. High-level functions are thus available
for both Trace.Exp.t and State.Exp.t